backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
mangum>=0.17.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
boto3>=1.29.7
openai>=1.0.0
python-docx>=1.0.0
pypdf>=3.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

